backend/requirements.txt
# Django and DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
django-extensions==3.2.3

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# API Documentation
drf-spectacular==0.27.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Storage
django-storages==1.14.2
boto3==1.34.0
Pillow==10.2.0

# PDF Generation
WeasyPrint==60.2
reportlab==4.0.8

# Audit & History
django-simple-history==3.4.0

# Multi-tenancy (choose one)
django-tenants==3.6.1
# OR for RLS approach:
# django-rls==0.1.0

# Internationalization
django-modeltranslation==0.18.11

# Validation & Serialization
marshmallow==3.20.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
urllib3==2.1.0

# Development & Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
bandit==1.7.5

# Logging
structlog==23.2.0

# Monitoring
sentry-sdk==1.39.2

# Excel/CSV handling
openpyxl==3.1.2
pandas==2.1.4

# Cryptography
cryptography==41.0.8

# HTTP client
httpx==0.26.0

# Time zone handling
zoneinfo==0.2.1

# UUID utilities
shortuuid==1.0.11

# Decimal handling
decimal==1.70

# JSON handling
orjson==3.9.10

# Environment detection
django-debug-toolbar==4.2.0

# Health checks
django-health-check==3.17.0

# Rate limiting
django-ratelimit==4.1.0

# CORS
django-cors-headers==4.3.1

# Security
django-csp==3.7
django-permissions-policy==4.17.0

# Background tasks
django-background-tasks==1.2.8

# Email
django-anymail==10.2

# Forms
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Admin enhancements
django-admin-interface==0.26.1
django-colorfield==0.10.1

# API versioning
django-rest-framework-api-key==3.0.0

# Caching
django-redis==5.4.0

# Search
django-haystack==3.2.1

# Backup
django-dbbackup==4.0.2
